modify makefile

+++ Modules/StepMotor/StepMotor.h
@@ -36,10 +36,10 @@ typedef struct
     Control_Mode_e control_mode;    // 控制模式: 0x01=速度控制, 0x02=位置控制, 0x03=力矩控制, 0x04=单圈绝对角度控制
     Motor_Direction_e motor_direction; // 转向: 0=逆时针, 1=顺时针
     uint8_t subdivision;     // 细分值: 例如0x20=32细分
-    uint8_t data;            // 速度控制模式下为 0；
+    uint16_t data;            // 速度控制模式下为 0；
                              // 位置控制模式下的角度数据，单位为角度;
                              // 力矩控制模式下的电流大小数据，单位为毫安
-    uint8_t speed;           // 速度,单位为 rad/s
+    uint16_t speed;           // 速度,单位为 rad/s
     UART_HandleTypeDef* usart_handle; // 使用具体的 UART_HandleTypeDef 实例
     MotorControlFunction control;  // 电机控制函数句柄
 } StepMotor_Init_Config_s;
@@ -54,10 +54,10 @@ typedef struct
     Control_Mode_e control_mode;    // 控制模式: 0x01=速度控制, 0x02=位置控制, 0x03=力矩控制, 0x04=单圈绝对角度控制
     Motor_Direction_e motor_direction; // 转向: 0=逆时针, 1=顺时针
     uint8_t subdivision;     // 细分值: 例如0x20=32细分
-    uint8_t data;            // 速度控制模式下为 0；
+    uint16_t data;            // 速度控制模式下为 0；
                              // 位置控制模式下的角度数据，单位为角度;
                              // 力矩控制模式下的电流大小数据，单位为毫安
-    uint8_t speed;           // 速度,单位为 rad/s
+    uint16_t speed;           // 速度,单位为 rad/s
     UART_HandleTypeDef* usart_handle; // 使用具体的 UART_HandleTypeDef 实例
     USARTInstance* usart_instance; // 添加指向 USART 实例的指针
     MotorControlFunction control;  // 电机控制函数句柄
